feat: add github OAuth

+++ backend/src/modules/users/users.service.ts
@@ -2,7 +2,7 @@ import { User } from '@prisma/client';
 
 import { ApiError } from '@/common/errors/apiError';
 
-import { RegisterPayload } from '../auth/auth.schema';
+import { RegisterPayload, RegisterWithOAuthPayload } from '../auth/auth.schema';
 
 import { UsersRepository } from './users.repository';
 import { UpdateUserPayload } from './users.schema';
@@ -14,41 +14,53 @@ export class UserService {
     this.usersRepository = userRepository;
   }
 
-  async getUserById(id: string): Promise<User> {
-    const user = await this.usersRepository.findById(id);
-    if (!user) {
-      throw ApiError.notFound('User not found');
-    }
-    return user;
-  }
-
-  async getUserByEmail(email: string): Promise<User> {
-    const user = await this.usersRepository.findByEmail(email);
-    if (!user) {
-      throw ApiError.notFound('User not found');
-    }
-    return user;
-  }
-
-  async createUser(payload: RegisterPayload): Promise<User> {
+  async createUser(
+    payload: RegisterPayload | RegisterWithOAuthPayload,
+  ): Promise<User> {
     const existingUser = await this.usersRepository.findByEmail(payload.email);
 
     if (existingUser) {
-      throw ApiError.badRequest('User with this email already exists', [
-        'Email already taken',
-      ]);
+      throw ApiError.badRequest('User with this email already exists');
     }
 
     return await this.usersRepository.create(payload);
   }
 
   async updateUser(id: string, payload: UpdateUserPayload): Promise<User> {
-    await this.getUserById(id);
+    const existingUser = await this.getUserById(id);
+
+    if (!existingUser) {
+      throw ApiError.notFound('User not found');
+    }
+
     return await this.usersRepository.update(id, payload);
   }
 
   async deleteUser(id: string): Promise<void> {
-    await this.getUserById(id);
+    const existingUser = await this.getUserById(id);
+
+    if (!existingUser) {
+      throw ApiError.notFound('User not found');
+    }
+
     await this.usersRepository.delete(id);
+  }
+
+  async getUserById(id: string): Promise<User | null> {
+    return this.usersRepository.findById(id);
+  }
+
+  async getUserByEmail(email: string): Promise<User | null> {
+    return this.usersRepository.findByEmail(email);
+  }
+
+  async getUserByProvider({
+    provider,
+    providerAccountId,
+  }: {
+    provider: string;
+    providerAccountId: string;
+  }): Promise<User | null> {
+    return this.usersRepository.findByProvider(provider, providerAccountId);
   }
 }
